requirements.txt
asgiref==3.11.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
Pillow==10.0.1
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
